refactor: column is no longer a cell
